feat: implemented a basic logging system
Implemented a basic logging system for use during the intial develoment of the tools and libraries. Logging is a achieved via a ILogAppender, currently only one is implemented the ConsoleAppender.

There is also support for a log dump from an internal StringBuilder in the event that no logging is configured but the tool crashes.