feat(deployment): setup DacTools.Deployment project
This project will eventually be the EXE as well as the dotnet tool for running deployments. This project uses Microsoft.Extensions.Hosting and a couple of the components of DacTools.Deployment.Core were reworked to support this as we now have DependencyInjection available.

The CI build pipeline was also changed as the EXE supports .NET Framework 472, but this is only built/tested on Windows agents.