build: updated cake scripts to include pack steps
The new pack steps include creating a single file exe for CoreFxVersion30 for each native runtime as well as framework dependent executables for all 3 frameworks or just the two versions of core if the pipeline is not running on Windows.

There is also a task to GZip the native runtime executables into a format that will eventually be added to the build artifacts.

Also added extra logging and parameters to detect the branch the build is running against.