test(deployment): added tests for AsyncTaskRunner<T>
Added an extremely basic set of tests for AsyncTaskRunner, these ensure that all tasks are queued and they are all completed. In the future, we might want these test tasks to perform some work or perhaps run a Task.Delay to ensure it does infact take the time and do all of the work in order or in parallel depending on the settings.