feat(deployment): add progress, warning and error build server messages
Using the groundwork already in place with `IBuildServer` and `IBuildServerResolver` this commit adds;
progress, warning and error build server integration messages when running with a compatible build server.

When using a compatible build server such as AzurePipelines specially formatted log messages will be output
which can be read by the build server to trigger events or show additional information. For example, the
progress message that has been implemented will display the current progress of the deployment task queue.

This partially implements the requirements for issue #6